Conflicts:
	src/java/com/icp/sigipro/bodegas/dao/InventarioDAO.java
	src/java/com/icp/sigipro/bodegas/modelos/Inventario.java

+++ web/Bodegas/ActivosFijos/Ver.jsp
@@ -24,7 +24,7 @@
             <li> 
               <a href="/SIGIPRO/Bodegas/ActivosFijos?">Activos Fijos</a>
             </li>
-            <li class="active"> ${activofijo.getPlaca()} </li>
+            <li class="active"> ${activofijo.getNombreFormato()} </li>
           </ul>
         </div>
         <div class="col-md-8 ">
@@ -39,7 +39,7 @@
           <!-- COLUMN FILTER DATA TABLE -->
           <div class="widget widget-table">
             <div class="widget-header">
-              <h3><i class="fa fa-barcode"></i> ${activofijo.getPlaca()} </h3>
+              <h3><i class="fa fa-barcode"></i> ${activofijo.getNombreFormato()} </h3>
               <div class="btn-group widget-header-toolbar">
                 <c:set var="contienePermisoEliminar" value="false" />
                 <c:forEach var="permiso" items="${sessionScope.listaPermisos}">
@@ -48,7 +48,7 @@
                   </c:if>
                 </c:forEach>
                 <c:if test="${contienePermisoEliminar}">
-                  <a class="btn btn-danger btn-sm boton-accion" href="/SIGIPRO/Bodegas/ActivosFijos?accion=eliminar&id_activo_fijo=${activofijo.getId_activo_fijo()}">Eliminar</a>
+                  <a class="btn btn-danger btn-sm boton-accion confirmable" data-texto-confirmacion="el activo fijo" data-href="/SIGIPRO/Bodegas/ActivosFijos?accion=eliminar&id_activo_fijo=${activofijo.getId_activo_fijo()}">Eliminar</a>
                 </c:if>
 
                 <c:set var="contienePermisoEditar" value="false" />
